[refactor][FIB-571] drop the AutoLamellaUI re-export from the package __init__
fibsem/applications/autolamella/ui/__init__.py existed only to re-export
AutoLamellaUI. FIB-559 had to make that export lazy (PEP 562) to break an import
loop; removing it outright is the better fix, and takes the lazy hook with it.

The export was serving ten TYPE_CHECKING blocks and two modules in
workflows/legacy/ -- so ten of the twelve edits here cannot affect runtime
behaviour at all. Both legacy modules already fail to import on main for an
unrelated reason (AutoLamellaProtocol no longer exists in structures.py), so
nothing live consumed the package-level name.

The try/except that wrapped it was migration scaffolding from c6576520
[refactor] Migrate autolamella into repo (#11). Nothing anywhere checked
`AutoLamellaUI is None`, so the dummy could never be observed by a caller -- it
only converted a real ImportError into a confusing one. During FIB-559 it hid a
genuine circular import for most of an afternoon: AutoLamellaUI silently became
None, the application still launched, and every moved module imported cleanly.

Importing anything from this package no longer runs the whole application's
import first, so the loop cannot re-form when future widgets move in.

Verified: 38 entry points cold-imported in fresh processes, each asserting the
class still resolves from its own module; the package attribute now raises
AttributeError rather than falling back to None.
