+++ fibsem/applications/autolamella/workflows/core.py
@@ -19,7 +19,7 @@ from fibsem.structures import (
 )
 
 if TYPE_CHECKING:
-    from fibsem.applications.autolamella.ui import AutoLamellaUI
+    from fibsem.applications.autolamella.ui.AutoLamellaUI import AutoLamellaUI
     from fibsem.applications.autolamella.structures import Lamella
 
 # CORE WORKFLOW STEPS
